Moved code into separate files. First step to de-monolithizationing.
