+++ .gitignore
@@ -1,1 +1,2 @@
-Dragon Warrior Monsters 2 Randomizer/dbsettings.php
+Dragon Warrior Monsters 2 Randomizer/config/dbsettings.php
+Dragon Warrior Monsters 2 Randomizer/config/settings.php
