+++ Dragon Warrior Monsters 2 Randomizer/config/.htaccess
@@ -1,0 +1,1 @@
+Deny from all
