+++ Dragon Warrior Monsters 2 Randomizer/include/.htaccess
@@ -1,0 +1,1 @@
+Deny from all
