+++ Dragon Warrior Monsters 2 Randomizer/index.php
@@ -1,1776 +1,29 @@
 <?php
-$dbaddress = 'localhost';
-$dbuser = 'USERNAME';
-$dbpass = 'PASSWORD';
-$dbname = 'dragonquestmonsters';
-
-// By default, require the user to upload a ROM themselves, and serve the patched result as a
-// downloadable file. Also, choose default location and filenames for processing ROMs server-side.
-$loadLocalRom = false;
-$saveLocalRom = false;
-$localRomDirectory = "F:\\ROMs\\Gameboy\\DWM2TA\\";
-$localRomInputName = "DWM2TA.gbc";
-$localRomOutputName = "DWM2TA_Random.gbc";
-
-// dbsettings.php overwrites the four $db variables, forces SSL, and turns off error reporting
-// ...and possibly overwrites other things from above. I don't know. Do what you want in there.
-require_once('dbsettings.php');
-
-$db = mysqli_connect($dbaddress,$dbuser,$dbpass);
-mysqli_select_db($db,$dbname);
-$dbreturn = '';
-//echo 'Database initialized.<br>';
-
-
-//Sends a query to the database.  Return value is stored in dbreturn.
-function execute($query){
-	global $dbreturn;
-	global $db;
-	$dbreturn = mysqli_query($db,$query);
-	if(!$dbreturn){
-		error_log("Error with query: ".$query."\n<br>".mysqli_error($db).'\n<br>');
-		echo mysqli_error($db);
-	}
-}
-//Translate dbreturn into a readable PHP array
-function get(){
-	global $dbreturn;
-	global $db;
-	$return = array();
-	if($dbreturn !== FALSE){
-		$array=mysqli_fetch_array($dbreturn);
-		if(is_array($array)){
-			foreach($array as $key=>$index){
-				if(!is_numeric($key)){
-					$return[$key] = $array[$key];
-				}
-			}
-		}
-	}
-	return $return;
-}
-
-$initial_seed = 0;
-
-//I'm a fan of memes.  Are you a fan of memes?  Here's a good meme: This is Dragon Warrior III's random number generator.
-//Airkix please do not RNG manip my randomizer.
-$seed = 0;
-$counter = 0;
-$discard = 0;
-
-function Random($MultiRandom = true){ //Defaulting MultiRandom to true for this randomizer.
-	global $counter;
-	global $seed;
-	global $discard;
-	//If we're using the "Battle RNG" value, we throw away that many values before keeping one.
-	$discarded = 0; //Throw away this many values
-	if($MultiRandom == true){
-		$discarded = $discard + 1;
-	}
-	for($j = 0; $j < $discarded+1; $j++){
-		for ($i = 0; $i < 16; $i++){
-			$seed = (($seed << 1) % 65536) ^ ((($seed >> 15) ^ 1) ? 0x1021 : 0);
-			if($seed >= 65536) $seed = $seed % 65536;
-		}
-		$counter++;
-		if($counter >= 256) $counter = $counter % 256;
-	}
-	return ($counter + $seed) % 256;
-}
 
 // Site with useful information: https://dragon-warrior-monsters-2-modding.fandom.com/wiki/Dragon_Warrior_Monsters_2_Modding_HomePage
 
-$romData;
-$ValidMonsterIDs = array(); //This is the actual ID number of the monster
-$ValidMonsterGrowthIndecies = array(); //This is the position of the monster in the "growths" list
-$MonsterNames = array(); // Monster ID -> Name
-$MonsterIDsByName = array(); // Monster Name -> ID
-$SkillNames = array(); // Skill ID -> Name
-$SkillIDsByName = array(); // Skill Name -> ID
-$encounter_data_length = 26;
-$first_bank_encounter_start = 0xD0075 + $encounter_data_length;
-$first_bank_encounter_count = 591;
-$second_bank_encounter_start = 0x288056 + $encounter_data_length;
-$second_bank_encounter_count = 93;
-$encounter_count = $first_bank_encounter_count + $second_bank_encounter_count;
-$monster_data_length = 47;
-$first_monster_byte = 0xD436A;
-$monster_count = 323;
-$item_strings_start = 0x224E3;
-$item_strings_count = 91;
-$item_behavior_start = 0x58CC2;
-$item_behavior_count = 99;
-$item_behavior_length = 13;
-
-$magicValues = array(
-	'HoodSquid Drop Encounter' => 26,
-
-	'Healing Skills' => array(
-		22,23,24,25,26, // Healing spells
-		30,31,32, // Vivify/Revive/Farewell
-		123,124, // LifeDance/Hustle
-		133, // TatsuCall (Tatsu can cast HealMore)
-		160,161, // LifeSong/LoveRain
-	),
-);
-$error = false;
-$error_message = 'The following errors occurred while generating the new seed:';
-
-
-$FlagSettings = array(
-	'Growth' => array(
-		'default' => 'None',
-		'label' => 'Monster Growth',
-		'options' => array(
-			array('value' => 'Redistribute', 'description' => 'Monster stat growth values will add to the same total, but will be randomly distributed.'),
-			array('value' => 'Shuffle', 'description' => 'Monsters will keep the same six growth values, but they will be randomly shuffled.'),
-			array('value' => 'None', 'description' => 'Do not randomize monster stats.'),
-		),
-	),
-
-	'Resistance' => array(
-		'default' => 'None',
-		'label' => 'Monster Resistances',
-		'options' => array(
-			array('value' => 'Redistribute', 'description' => 'Monster stat resistance values will add to the same total, but will be randomly distributed.'),
-			array('value' => 'Shuffle', 'description' => 'Monsters will keep the same 27 resistance values, but they will be randomly shuffled.'),
-			array('value' => 'None', 'description' => 'Do not randomize monster resistances.'),
-		),
-	),
-
-	'Skills' => array(
-		'default' => 'None',
-		'label' => 'Shuffle Monster Skills',
-		'options' => array(
-			array('value' => 'Random', 'description' => 'Monsters learn completely random skills. BeDragon is excluded.'),
-			array('value' => 'Random With BeDragon', 'label' => 'Random With BeDragon <b>(!)</b>', 'description' => 'Monsters learn completely random skills, including BeDragon. Boo!'),
-			array('value' => 'Random, No Healing', 'label' => 'Random, No Healing <b>(!)</b>', 'description' => 'Monsters learn completely random skills. BeDragon and all skills with healing capabilities are excluded.'),
-			array('value' => 'None', 'description' => 'Do not randomize monster skills.'),
-		),
-	),
-
-	'Encounters' => array(
-		'default' => 'None',
-		'label' => 'Shuffle Encounters',
-		'options' => array(
-			array('value' => 'Poorly', 'description' => 'Shuffle enemy types. Monsters\' stats add to the same total, proportional to their growth values. Later monsters are supposed to start with more skills, but no monsters start with advanced skills.'),
-			array('value' => 'None', 'description' => 'Do not randomize encounters.'),
-		),
-	),
-
-	'GeniusMode' => array(
-		'default' => 'Off',
-		'label' => 'Max Monster Intelligence',
-		'options' => array(
-			array('value' => 'On', 'description' => 'All wild monsters have 999 Int; all monsters have 31 int growth. (Overrides randomized base int/growth)'),
-			array('value' => 'Off', 'description' => 'Do not max out monster intelligence.'),
-		),
-	),
-
-	'EXPScaling' => array(
-		'default' => '100',
-		'label' => 'Experience Yields',
-		'options' => array(
-			array('value' => '0', 'label' => '0%', 'description' => 'Monsters will be worth 0% of their normal EXP yields (minimum 1 EXP)'),
-			array('value' => '50', 'label' => '50%', 'description' => 'Monsters will be worth 50% of their normal EXP yields (minimum 1 EXP)'),
-			array('value' => '100', 'label' => '100%', 'description' => 'Monsters will be worth 100% of their normal EXP yields'),
-			array('value' => '200', 'label' => '200%', 'description' => 'Monsters will be worth 200% of their normal EXP yields'),
-			array('value' => '500', 'label' => '500%', 'description' => 'Monsters will be worth 500% of their normal EXP yields'),
-		),
-	),
-
-	'StatScaling' => array(
-		'default' => '100',
-		'label' => 'Wild Monster Stat Scaling',
-		'options' => array(
-			array('value' => '0', 'label' => '0%', 'description' => 'Wild monster stats will be altered based on this scalar'),
-			array('value' => '50', 'label' => '50%', 'description' => 'Wild monster stats will be altered based on this scalar'),
-			array('value' => '100', 'label' => '100%', 'description' => 'Wild monster stats will be altered based on this scalar'),
-			array('value' => '200', 'label' => '200%', 'description' => 'Wild monster stats will be altered based on this scalar'),
-			array('value' => '500', 'label' => '500%', 'description' => 'Wild monster stats will be altered based on this scalar'),
-		),
-	),
-
-	'BossScaling' => array(
-		'default' => '100',
-		'label' => 'Boss Monster Stat Scaling',
-		'options' => array(
-			array('value' => '0', 'label' => '0%', 'description' => 'Boss stats will be altered based on this scalar'),
-			array('value' => '50', 'label' => '50%', 'description' => 'Boss stats will be altered based on this scalar'),
-			array('value' => '100', 'label' => '100%', 'description' => 'Boss stats will be altered based on this scalar'),
-			array('value' => '200', 'label' => '200%', 'description' => 'Boss stats will be altered based on this scalar'),
-			array('value' => '300', 'label' => '300%', 'description' => 'Boss stats will be altered based on this scalar'),
-			array('value' => '400', 'label' => '400%', 'description' => 'Boss stats will be altered based on this scalar'),
-			array('value' => '500', 'label' => '500%', 'description' => 'Boss stats will be altered based on this scalar'),
-		),
-	),
-
-	'YetiMode' => array(
-		'default' => 'Off',
-		'label' => 'Yeti Mode',
-		'options' => array(
-			array('value' => 'On', 'description' => 'All monsters are yetis!'),
-			array('value' => 'Off', 'description' => 'Not all monsters are yetis.'),
-		),
-	),
-);
-
-$Flags = array();
-
-function DWM2R()
-{
-	global $Flags;
-	global $FlagSettings;
-	global $initial_seed;
-	global $romData;
-	
-	//Parse out the flags.  I'm basing it on the individual form fields, not the flags string, because the flags string is dynamically generated on the JS front-end, and I can't really make that dynamic on the PHP side.
-	if(array_key_exists("StartingMonster",$_REQUEST)){
-		$Flags["StartingMonster"] = trim($_REQUEST["StartingMonster"]);
-	}else{
-		$Flags["StartingMonster"] = 0;
-	}
-	foreach ($FlagSettings as $name => $settings) {
-		if(array_key_exists("default", $settings)) {
-			if(array_key_exists($name,$_REQUEST)) {
-				$Flags[$name] = trim($_REQUEST[$name]);
-			}else{
-				$Flags[$name] = $settings["default"];
-			}
-		}
-	}
-	if(array_key_exists("Seed",$_REQUEST)){
-		$Flags["Seed"] = trim($_REQUEST["Seed"]);
-	}else{
-		$Flags["Seed"] = 0;
-	}
-	$initial_seed = $Flags["Seed"];
-	
-	if(array_key_exists("Submit",$_REQUEST)){
-		if (!loadRom())
-			return;
-
-		PopulateMetadata();
-		
-		//Some functions to dump the ROM in hexidecimal or text format
-		//RomStructuredDataDump();
-		//RomDump();
-		//RomTextDump();
-		//This function will attempt to find all of the bosses in the ROM data based on stats I pulled from Gamefaqs
-		//LocateBosses();
-		//die();
-		if(strlen($romData) > 0){
-			hackRom();
-			saveRom();
-		}
-	}
-}
-
-function hackRom()
-{
-	//This function checks the selected flags and determines which randomization subroutines to run
-	
-	//First, let's seed the random number generator.  We're using DW3's generator, which uses three variables.
-	global $counter;
-	global $seed;
-	global $discard;
-	global $initial_seed;
-	global $flags;
-	$tmp_seed = $initial_seed;
-	
-	$counter = $tmp_seed % 256;
-	$tmp_seed = floor($tmp_seed / 256);
-	$seed = $tmp_seed % 65536;
-	$tmp_seed = floor($tmp_seed / 65536);
-	$discard = $tmp_seed % 16;
-	
-	for($j = 0; $j < $flags["StartingMonster"]; $j++){
-		//Burn random numbers to make monster choice affect the randomization.
-		//TODO: Should all flags do this?
-		Random();
-	}
-	
-	//Now, we actually do the randomizing.  Each step of the randomizer is its own function, for code cleanliness.
-	ShuffleMonsterGrowth();
-	ShuffleMonsterResistances();
-	ShuffleMonsterSkills();
-	ShuffleEncounters();
-	//TODO: Shuffle fixed items
-	//TODO: Shuffle random items?
-	//TODO: Shuffle shops/item prices
-	CodePatches();
-	Analytics();
-	
-	return true;
-}
-
-
-function Analytics()
-{
-	//This function logs intrusive analytical data so I can see how often this randomizer gets played
-	global $initial_seed;
-	global $flags;
-	
-	//Logging IP address, user agent, flags, seed, and current time.
-	//Is your user agent any of my business?  Not really, but don't worry about it.
-	execute("
-		INSERT INTO dwm2r_activitymonitor
-		(IPAddress,UserAgent,Flags,Seed,CreatedDTS)
-		VALUES (
-			'".$_SERVER['REMOTE_ADDR']."',
-			'".$_SERVER['HTTP_USER_AGENT']."',
-			'".trim($_REQUEST["Flags"])."',
-			'".$initial_seed."',
-			NOW()
-		)
-	");
-}
-
-function loadRom()
-{
-	global $romData;
-	global $localRomDirectory;
-	global $localRomInputName;
-	global $loadLocalRom;
-	try
-	{
-		$filename = "";
-		if ($loadLocalRom) {
-			//This is code for processing a ROM that I already have on the server
-			$filename = $localRomDirectory.$localRomInputName;
-		} else {
-			$filename = $_FILES['InputFile']['tmp_name'];
-		}
-
-		$file = fopen($filename, "rb");
-		$romData = fread($file,filesize($filename));
-		
-		//When $romData is a standard PHP array, this 4MB file takes up over 128M of RAM.
-		//When $romData is an SplFixedArray, this 4MB file takes up 67MB of RAM.
-		//for($i = 0; $i < strlen($_romData); $i++){
-		//	$romData[$i] = ord($_romData[$i]);
-		//}
-		
-		fclose($file);
-	}
-	catch (Exception $e)
-	{
-		$error_message = "<br>Empty file name(s) or unable to open files.  Please verify the files exist.";
-		return false;
-	}
-	return true;
-}
-
-function saveRom()
-{
-	global $romData;
-	global $initial_seed;
-	global $localRomDirectory;
-	global $localRomOutputName;
-	global $saveLocalRom;
-	
-	if ($saveLocalRom) {
-		//This is code for saving a rom to the server instead of spitting it back out to the user
-		$filename = $localRomDirectory.$localRomOutputName;
-		$file = fopen($filename, "w");
-		fwrite($file,$romData);
-		fclose($file);
-	} else {
-		header('Content-Disposition: attachment; filename="DWM2_Rando_'.$initial_seed.'.gbc"');
-		header("Content-Size: ".strlen($romData)*512);
-		echo $romData;
-		die();
-	}
-}
-
-function swap($firstAddress, $secondAddress)
-{
-	//This function just switches one value in the rom data with another.  Useful when shuffling data.
-	global $romData;
-	$holdAddress = $romData[$secondAddress];
-	$romData[$secondAddress] = $romData[$firstAddress];
-	$romData[$firstAddress] = $holdAddress;
-}
-
-function getByte($offset) {
-	global $romData;
-	return ord($romData[$offset]);
-}
-
-function getWord($offset) {
-	global $romData;
-	return ord($romData[$offset]) + ord($romData[$offset + 1])*256;
-}
-
-function setByte($offset, $value) {
-	global $romData;
-	$romData[$offset] = chr(floor($value % 256));
-}
-
-function setWord($offset, $value) {
-	global $romData;
-	$romData[$offset] = chr(floor($value % 256));
-	$romData[$offset + 1] = chr(floor(($value / 256) % 256));
-}
-
-function calcMonsterOffset($i, $offset) {
-	global $first_monster_byte;
-	global $monster_data_length;
-	return $first_monster_byte + $i * $monster_data_length + $offset;
-}
-
-function getMonsterByte($i, $offset)
-{
-	return getByte(calcMonsterOffset($i, $offset));
-}
-
-function getMonsterWord($i, $offset)
-{
-	return getWord(calcMonsterOffset($i, $offset));
-}
-
-function setMonsterByte($i, $offset, $value)
-{
-	return setByte(calcMonsterOffset($i, $offset), $value);
-}
-
-function calcEncounterOffset($i, $offset) {
-	global $encounter_data_length;
-	global $first_bank_encounter_start;
-	global $first_bank_encounter_count;
-	global $second_bank_encounter_start;
-	
-	if ($i < $first_bank_encounter_count) {
-		return ($first_bank_encounter_start + $i * $encounter_data_length + $offset);
-	}
-	return ($second_bank_encounter_start + ($i - $first_bank_encounter_count) * $encounter_data_length + $offset);
-}
-
-function getEncounterByte($i, $offset)
-{
-	return getByte(calcEncounterOffset($i, $offset));
-}
-
-function getEncounterWord($i, $offset)
-{
-	return getWord(calcEncounterOffset($i, $offset));
-}
-
-function setEncounterByte($i, $offset, $value)
-{
-	return setByte(calcEncounterOffset($i, $offset), $value);
-}
-
-function setEncounterWord($i, $offset, $value)
-{
-	return setWord(calcEncounterOffset($i, $offset), $value);
-}
-
-function swapEncounterBytes($i, $offset_a, $offset_b)
-{
-	swap(calcEncounterOffset($i, $offset_a), calcEncounterOffset($i, $offset_b));
-}
-
-
-function WriteText($address, $text)
-{
-	//This function translates text to binary code.  Text is uncompressed in DWM2, but it isn't mapped the same as ASCII.
-	global $romData;
-	$i = 0;
-	for($j = 0; $j < strlen($text); $j++)
-	{
-		$c = $text[$j];
-		$x = 0;
-		if($c >= 'a' && $c <= 'z')
-		{
-			$x = ord($c) - ord('a') + 0x24;
-		}else if($c >= 'A' && $c <= 'Z')
-		{
-			$x = ord($c) - ord('A') + 0x0A;
-		}else if($c >= '0' && $c <= '9')
-		{
-			$x = ord($c) - '1';
-			if($c == '0')
-			{
-				$x += 10;
-			}
-		}
-		else
-		{
-			$x = 0x90;
-		}
-		$romData[$address + $i] = chr($x);
-		$i++;
-	}
-}
-
-function RomDump(){
-	/* Code to create a hexidecimal rom dump of the input rom. */
-	global $romData;
-	global $localRomDirectory;
-
-	$outfilename = $localRomDirectory.'romdump100_w_text.txt';
-	$outfile = fopen($outfilename, "w");
-	
-	$line = '0x000000  ';
-	$hex = "";
-	$str = "";
-	for($i = 0; $i < strlen($romData); $i++){
-		$byte = getByte($i);
-		$hex .= str_pad(dechex($byte), 2, '0', STR_PAD_LEFT) . ' ';
-		if ($byte >= 10 && $byte < 36) {
-			$str .= chr(ord('A') + ($byte - 10));
-		} elseif ($byte >= 36 && $byte < 62) {
-			$str .= chr(ord('a') + ($byte - 36));
-		} elseif ($byte == 0x90) {
-			$str .= ' ';
-		} elseif ($byte < 10) {
-			$str .= chr(ord('0') + $byte);
-		} else {
-			$str .= '.';
-		}
-		if($i % 64 == 63){
-			fwrite($outfile,$line . $str . '   ' . $hex."\n");
-			$line = '0x' . str_pad(dechex($i + 1), 6, '0', STR_PAD_LEFT) . '  ';
-			$hex = '';
-			$str = '';
-		}
-	}
-	fclose($outfile);
-}
-
-function RomTextDump(){
-	/* Code to create a text dump of the input rom. */
-	global $romData;
-	global $localRomDirectory;
-	
-	$files_to_create = 1;
-	
-	for($j = 0; $j < $files_to_create; $j++){
-		$outfilename = $localRomDirectory.'romtextdump'.$j.'.txt';
-		$outfile = fopen($outfilename, "w");
-		
-		$str = '';
-		for($i = strlen($romData)/$files_to_create*$j; $i < strlen($romData)/$files_to_create*($j+1); $i++){
-			if((ord($romData[$i]) >= 0x24) && (ord($romData[$i]) < (0x24 + 26))){
-				$str .= chr(ord($romData[$i]) - 0x24 + ord('a'));
-			}
-			elseif((ord($romData[$i]) >= 0x0A) && (ord($romData[$i]) < (0x0A + 26))){
-				$str .= chr(ord($romData[$i]) - 0x0A + ord('A'));
-			}
-			elseif((ord($romData[$i]) >= 0) && (ord($romData[$i]) < 0x10)){
-				$str .= chr(ord($romData[$i]) + ord('0'));
-			}
-			else{
-				$str .= ' ';
-			}
-			if($i % 100 == 99){
-				$str .= "\n";
-			}
-		}
-		fwrite($outfile,$str);
-		fclose($outfile);
-	}
-}
-
-function RomStructuredDataDump() {
-	RomEncounterDump();
-	RomMonsterDump();
-	RomItemStringsDump();
-	RomItemBehaviorDump();
-	RomBank33Dump();
-}
-
-function RomRawEncounterDump() {
-	global $romData;
-	global $encounter_data_length;
-	global $encounter_count;
-	global $localRomDirectory;
-
-	$outfilename = $localRomDirectory.'rom_raw_encounter_dump.txt';
-	$outfile = fopen($outfilename, "w");
-	for ($i = 0; $i < $encounter_count; $i++) {
-		$str = "";
-		for ($j = 0; $j < $encounter_data_length; $j++) {
-			$byte = getEncounterByte($i, $j);
-			$str .= str_pad(dechex($byte), 2, '0', STR_PAD_LEFT) . ' ';
-		}
-		fwrite($outfile,$str."\n");
-	}
-	fclose($outfile);
-}
-
-function RomEncounterDump() {
-	// Encounter data guide
-	// Offset:   Size:    Description:
-	// 0         2        Monster ID
-	// 2-5       4*1      Monster Skills
-	// 6         2        Monster EXP
-	// 8         1        Join Rate (Values from 0-7. 0 = max join chance, 7 = never)
-	// 9         1        Monster level
-	// 10        2        Monster HP
-	// 12        2        Monster MP
-	// 14        2        Monster ATK
-	// 16        2        Monster DEF
-	// 18        2        Monster AGL
-	// 20        2        Monster INT
-	// 22-25     4x1      Personality Traits / Motivation
-	global $romData;
-	global $encounter_data_length;
-	global $encounter_count;
-	global $localRomDirectory;
-	global $MonsterNames;
-
-	$outfilename = $localRomDirectory.'rom_encounter_dump.txt';
-	$outfile = fopen($outfilename, "w");
-	$str = "Row     mID Name       Boss |--Skills-|   EXP Jn LV    HP    MP   ATK   DEF   AGL   INT Personality ";
-	fwrite($outfile, $str."\n");
-	$str = "Offset:   0                  2  3  4  5   6-7  8  9 10-11 12-13 14-15 16-17 18-19 20-21 22 23 24 25 ";
-	fwrite($outfile, $str."\n");
-	$str = "--------------------------------------------------------------------------------------------------- ";
-	fwrite($outfile, $str."\n");
-	for ($i = 0; $i < $encounter_count; $i++) {
-		$str = "";
-		$str .= str_pad($i, 3, ' ', STR_PAD_LEFT) . '     ';
-		$monster_id = getEncounterWord($i, 0);
-		$str .= str_pad($monster_id, 3, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad($MonsterNames[$monster_id], 10, ' ') . ' ';
-		$str .= (isArenaEncounter($i) ? 'Arna' : (isBossEncounter($i) ? 'Boss' : (isBossRecruit($i) ? 'Recr' : '    '))) . ' ';
-		for ($j = 2; $j <= 5; $j++) {
-			$str .= str_pad(dechex(getEncounterByte($i, $j)), 2, '0', STR_PAD_LEFT) . ' ';
-		}
-		$str .= str_pad(getEncounterWord($i, 6), 5, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(dechex(getEncounterByte($i, 8)), 2, '0', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(getEncounterByte($i, 9), 2, ' ', STR_PAD_LEFT) . ' ';
-		for ($j = 10; $j <= 20; $j += 2) {
-			$str .= str_pad(getEncounterWord($i, $j), 5, ' ', STR_PAD_LEFT) . ' ';
-		}
-		for ($j = 22; $j <= 25; $j++) {
-			$str .= str_pad(dechex(getEncounterByte($i, $j)), 2, '0', STR_PAD_LEFT) . ' ';
-		}
-		fwrite($outfile, $str."\n");
-	}
-	fclose($outfile);
-}
-
-function RomRawMonsterDump() {
-	global $romData;
-	global $monster_data_length;
-	global $monster_count;
-	global $localRomDirectory;
-
-	$outfilename = $localRomDirectory.'rom_monster_dump.txt';
-	$outfile = fopen($outfilename, "w");
-	for ($i = 0; $i < $monster_count; $i++) {
-		$str = "";
-		for ($j = 0; $j < $monster_data_length; $j++) {
-			$byte = getMonsterByte($i, $j);
-			$str .= str_pad(dechex($byte), 2, '0', STR_PAD_LEFT) . ' ';
-		}
-		fwrite($outfile,$str."\n");
-	}
-	fclose($outfile);
-}
-
-function RomMonsterDump() {
-	// Monster data guide
-	// Offset:   Size:    Description:
-	// 0         1        Monster family (00-0a)
-	// 1         1        Sex distribution (Possible values are 0-3, most are 2.)
-	// 2         1        Flying flag (e.g. LegSweep miss)
-	// 3         1        Metal / coward flag (Metaly, Metabble, and MetalKing have 1, the rest are 0.)
-	// 4         1        Join rate in random key worlds (Possible values are 1-7. 7 = never)
-	// 5         1        ??? (Possible values are 0, 1, 2, and 3. Almost always matches offset 2, except for monster family 0x09 and a couple other exceptions: GigaDraco: 1 here, but 0 in offset 2. PomPomBom: 0 here, but 1 in offset 2. Water family is almost always 2, except for Starfish, which is 3. Geyser / water attack related?)
-	// 6         1        Max unbred level
-	// 7         1        Exp required for next level growth
-	// 8-10      3*1      Learnable Skills
-	// 11        1        ??? (observed values are 0, 1, 2, 3, and 5. Most are 0. Choices for rando world guardians?)
-	// 12-17     6*1      HP/MP/ATK/DEF/AGL/INT Growth
-	// 18-44     27*1     Resistances (0 = none, 1 = slight, 2 = some, 3 = immune)
-	// 45-46     2        Base EXP value on kill for random key worlds
-	global $romData;
-	global $monster_data_length;
-	global $monster_count;
-	global $localRomDirectory;
-	global $ValidMonsterGrowthIndecies;
-	global $MonsterNames;
-
-	$outfilename = $localRomDirectory.'rom_monster_dump.txt';
-	$outfile = fopen($outfilename, "w");
-	$str = "                                                   Gro  Skills      |----Growth-----| |-------Resistances-------|        ";
-	fwrite($outfile, $str."\n");
-	$str = "Row     mID Name       Group  ? Fly Run  ?  ? mLVL EXP |------|  ?  HP MP AT DF AG IN |-------------------------| KillXP ";
-	fwrite($outfile, $str."\n");
-	$str = "Offset:                    0  1   2   3  4  5    6   7  8  9 10 11  12 13 14 15 16 17 18-44                        45-46 ";
-	fwrite($outfile, $str."\n");
-	$str = "------------------------------------------------------------------------------------------------------------------------ ";
-	fwrite($outfile, $str."\n");
-	for ($i = 0; $i < $monster_count; $i++) {
-		$str = "";
-		$str .= str_pad($i, 7, ' ', STR_PAD_LEFT) . ' ';
-		$monster_id = $ValidMonsterGrowthIndecies[$i];
-		$str .= str_pad($monster_id, 3, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad($MonsterNames[$monster_id], 10, ' ') . ' ';
-		$str .= str_pad(getMonsterByte($i, 0), 5, ' ', STR_PAD_LEFT) . ' '; // Group
-		$str .= str_pad(dechex(getMonsterByte($i, 1)), 2, '0', STR_PAD_LEFT) . ' ';
-		$str .= (getMonsterByte($i, 2) == 1 ? "Yes" : "  .") . ' '; // Fly?
-		$str .= (getMonsterByte($i, 3) == 1 ? "Yes" : "  .") . ' '; // Run?
-		$str .= str_pad(dechex(getMonsterByte($i, 4)), 2, '0', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(dechex(getMonsterByte($i, 5)), 2, '0', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(getMonsterByte($i, 6), 4, ' ', STR_PAD_LEFT) . ' '; // Max LVL
-		$str .= str_pad(getMonsterByte($i, 7), 3, ' ', STR_PAD_LEFT) . ' '; // Next level EXP growth
-		for ($j = 8; $j <= 10; $j++) {
-			$str .= str_pad(dechex(getMonsterByte($i, $j)), 2, '0', STR_PAD_LEFT) . ' '; // Skills
-		}
-		$str .= str_pad(dechex(getMonsterByte($i, 11)), 2, '0', STR_PAD_LEFT) . ' ';
-		$str .= ' ';
-		for ($j = 12; $j <= 17; $j++) {
-			$str .= str_pad(getMonsterByte($i, $j), 2, ' ', STR_PAD_LEFT) . ' '; // Stat growth
-		}
-		for ($j = 18; $j <= 44; $j++) {
-			$res = getMonsterByte($i, $j);
-			if ($res == 0) {
-				$str .= ".";
-			} elseif ($res == 1) {
-				$str .= "o";
-			} elseif ($res == 2) {
-				$str .= "O";
-			} else {
-				$str .= "X";
-			}
-		}
-		$str .= ' ';
-		$str .= str_pad(getMonsterWord($i, 45), 6, ' ', STR_PAD_LEFT) . ' '; // Base EXP gain for kill?
-
-		fwrite($outfile,$str."\n");
-	}
-	fclose($outfile);
-}
-
-function RomItemStringsDump() {
-	global $romData;
-	global $item_strings_start;
-	global $item_strings_count;
-	global $localRomDirectory;
-	
-	$outfilename = $localRomDirectory.'rom_strings_dump.txt';
-	$outfile = fopen($outfilename, "w");
-	$str = "";
-	$hex = "";
-	$counter = 0;
-	for ($i = 0; $i < 1000; $i++) {
-		$byte = ord($romData[$item_strings_start + $i]);
-		$hex .= str_pad(dechex($byte), 2, '0', STR_PAD_LEFT) . ' ';
-		if ($byte == 0xf0) {
-			fwrite($outfile,str_pad($str, 18, ' ') . $hex . "\n");
-			$str = "";
-			$hex = "";
-			$counter += 1;
-			if ($counter >= $item_strings_count) {
-				break;
-			}
-		} elseif ($byte >= 10 && $byte < 36) {
-			$str .= chr(ord('A') + ($byte - 10));
-		} elseif ($byte >= 36 && $byte < 62) {
-			$str .= chr(ord('a') + ($byte - 36));
-		} elseif ($byte < 10) {
-			$str .= chr(ord('0') + $byte);
-		} else {
-			$str .= '.';
-		}
-	}
-	fclose($outfile);
-}
-
-function RomItemBehaviorDump() {
-	global $item_behavior_start;
-	global $item_behavior_count;
-	global $item_behavior_length;
-	global $localRomDirectory;
-	
-	$outfilename = $localRomDirectory.'rom_items_dump.txt';
-	$outfile = fopen($outfilename, "w");
-	$str = "Row     Type Price  ? Use  ? Target Icon  ? F1 F2 V1 V2 ";
-	fwrite($outfile,$str . "\n");
-	$str = "Offset:    0   1-2  3   4  5      6    7  8  9 10 11 12 ";
-	fwrite($outfile,$str . "\n");
-	$str = "------------------------------------------------------- ";
-	fwrite($outfile,$str . "\n");
-	
-	for ($i = 0; $i < $item_behavior_count; $i++) {
-		$str = "";
-		$str .= str_pad($i, 6, ' ', STR_PAD_LEFT) . '  ';
-		$str .= str_pad(getByte($item_behavior_start + $i * $item_behavior_length + 0), 4, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(getWord($item_behavior_start + $i * $item_behavior_length + 1), 5, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(dechex(getByte($item_behavior_start + $i * $item_behavior_length + 3)), 2, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(dechex(getByte($item_behavior_start + $i * $item_behavior_length + 4)), 3, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(dechex(getByte($item_behavior_start + $i * $item_behavior_length + 5)), 2, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(dechex(getByte($item_behavior_start + $i * $item_behavior_length + 6)), 6, ' ', STR_PAD_LEFT) . ' ';
-		$str .= str_pad(dechex(getByte($item_behavior_start + $i * $item_behavior_length + 7)), 4, ' ', STR_PAD_LEFT) . ' ';
-		for ($j = 8; $j <= 12; $j++) {
-			$str .= str_pad(dechex(getByte($item_behavior_start + $i * $item_behavior_length + $j)), 2, '0', STR_PAD_LEFT) . ' ';
-		}
-		fwrite($outfile,$str . "\n");
-	}
-	fclose($outfile);
-}
-
-function RomDumpStrings($outfile, $start, $end) {
-	$str = '';
-	for ($i = $start; $i <= $end; $i++) {
-		$byte = getByte($i);
-		if ($byte == 0xf0) {
-			fwrite($outfile, $str . "\n");
-			$str = '';
-		} elseif ($byte >= 10 && $byte < 36) {
-			$str .= chr(ord('A') + ($byte - 10));
-		} elseif ($byte >= 36 && $byte < 62) {
-			$str .= chr(ord('a') + ($byte - 36));
-		} elseif ($byte == 0x90) {
-			$str .= ' ';
-		} elseif ($byte < 10) {
-			$str .= chr(ord('0') + $byte);
-		} else {
-			$str .= '.';
-		}
-	}
-}
-
-function RomBank33Dump() {
-	global $localRomDirectory;
-	$outfilename = $localRomDirectory.'rom_bank_33_dump.txt';
-	$outfile = fopen($outfilename, "w");
-	fwrite($outfile, "Bank 0x33 (0x0CC000 - 0x0CFFFF): Dialog box / window structure, arena text\n\n");
-	fwrite($outfile, "Dialog box / window structures\n");
-	$counter = 0;
-	$i = 0xCC0C5;
-	while ($i <= 0xCD1C5) {
-		$str = "Offset " . $counter . ":\n";
-		$flag1 = getByte($i); $str .= str_pad(dechex($flag1), 2, '0', STR_PAD_LEFT) . ' '; $i += 1;
-		$flag2 = getByte($i); $str .= str_pad(dechex($flag2), 2, '0', STR_PAD_LEFT) . ' '; $i += 1;
-		$width = getByte($i); $str .= str_pad($width, 2, ' ', STR_PAD_LEFT) . ' '; $i += 1;
-		$height = getByte($i); $str .= str_pad($height, 2, ' ', STR_PAD_LEFT) . ' '; $i += 1;
-		fwrite($outfile, $str."\n");
-		for ($j = 0; $j < $height; $j++) {
-			$str = "";
-			for ($k = 0; $k < $width; $k++) {
-				$str .= str_pad(dechex(getByte($i + $j * $width + $k)), 2, '0', STR_PAD_LEFT) . ' ';
-			}
-			fwrite($outfile, $str."\n");
-		}
-		$i += $height * $width;
-		$counter += 1;
-		fwrite($outfile, "\n");
-	}
-	
-	fwrite($outfile, "\n\nArena Attendant strings:\n");
-	RomDumpStrings($outfile, 0xCE515, 0xCF036);
-	fwrite($outfile, "\n\nPvP combat / trade strings:\n");
-	RomDumpStrings($outfile, 0xCF074, 0xCF1EF);
-	fclose($outfile);
-}
-
-
-function PopulateMetadata(){
-	global $Flags;
-	global $ValidMonsterIDs;
-	global $ValidMonsterGrowthIndecies;
-	global $MonsterNames;
-	global $MonsterIDsByName;
-	global $SkillNames;
-	global $SkillIDsByName;
-	global $magicValues;
-
-	$monster_list_query = "SELECT * FROM dragonwarriormonsters2 order by id asc";
-	execute($monster_list_query);
-	while($monster = get()){
-		$MonsterNames[$monster["id"]] = $monster["name"];
-		$MonsterIDsByName[$monster["name"]] = $monster["id"];
-	}
-
-	$skill_list_query = "SELECT * FROM dragonwarriormonsters2_skills order by id asc";
-	execute($skill_list_query);
-	while($skill = get()){
-		$SkillNames[$skill["id"]] = $skill["Name"];
-		$SkillIdsByName[$skill["Name"]] = $skill["id"];
-	}
-	
-	if($Flags["YetiMode"] == "On"){
-		$ValidMonsterIDs[] = $MonsterIDsByName["Yeti"];
-	}
-	else
-	{
-		//This is the ID stored in the SRAM that determines which monster you have.
-		//It's also used within the table of base-stats for each monster.
-		//NOTE 0x1B is Butch and I don't think he should be used?
-		for ($i = 0; $i <= 0x17E; $i++)
-		{
-			if (
-				($i >= 0x01 && $i <= 0x1A) || //Slimes
-				($i >= 0x24 && $i <= 0x42) || //Dragons
-				($i >= 0x47 && $i <= 0x66) || //Beasts
-				($i >= 0x6A && $i <= 0x84) || //Birds
-				($i >= 0x8D && $i <= 0xA7) || //Plants
-				($i >= 0xB0 && $i <= 0xC9) || //Bugs
-				($i >= 0xD3 && $i <= 0xF0) || //Devils
-				($i >= 0xF6 && $i <= 0x110) || //Zombies
-				($i >= 0x119 && $i <= 0x138) || //Materials
-				($i >= 0x13C && $i <= 0x15B) || //Waters
-				($i >= 0x15F && $i <= 0x174) //Bosses
-				)
-			{
-				$ValidMonsterIDs[] = $i;
-			}
-		}
-	}
-	for ($i = 0; $i <= 0x17E; $i++)
-	{
-		if (
-			($i >= 0x01 && $i <= 0x1B) || //Slimes (0x1B is Butch)
-			($i >= 0x24 && $i <= 0x42) || //Dragons
-			($i >= 0x47 && $i <= 0x66) || //Beasts
-			($i >= 0x6A && $i <= 0x84) || //Birds
-			($i >= 0x8D && $i <= 0xA7) || //Plants
-			($i >= 0xB0 && $i <= 0xC9) || //Bugs
-			($i >= 0xD3 && $i <= 0xF0) || //Devils
-			($i >= 0xF6 && $i <= 0x110) || //Zombies
-			($i >= 0x119 && $i <= 0x138) || //Materials
-			($i >= 0x13C && $i <= 0x15B) || //Waters
-			($i >= 0x15F && $i <= 0x174) //Bosses
-			)
-		{
-			$ValidMonsterGrowthIndecies[] = $i;
-		}
-	}
-}
-
-
-function ShuffleMonsterGrowth()
-{
-	global $Flags;
-	
-	global $romData;
-	global $monster_count;
-
-	for ($i = 0; $i < $monster_count; $i++)
-	{
-		if ($Flags["Growth"] == "Redistribute")
-		{
-			//Let's randomize the monster's growth stats, but have them add up to the same value.
-			$total_stats = 0;
-			for ($j = 0; $j < 6; $j++)
-			{
-				//We're going to set a minimum growth value at 1 because growth of 0 SUCKS
-				$total_stats += getMonsterByte($i, 12 + $j) - 1;
-				setMonsterByte($i, 12 + $j, 1);
-			}
-
-			//Start by assigning 30 points: 20 to one stat and 10 to another (Or the same?)
-			$slot1 = Random() % 6; //Named slot1 because C# is throwing a fit if I re-use the same var name in the loop below...
-			setMonsterByte($i, 12 + $slot1, getMonsterByte($i, 12 + $slot1) + 20);
-			$slot1 = Random() % 6;
-			setMonsterByte($i, 12 + $slot1, getMonsterByte($i, 12 + $slot1) + 10);
-			$total_stats -= 30;
-
-			if($total_stats > 31*6) $total_stats = 31*6;
-			while($total_stats > 0)
-			{
-				$slot = Random() % 6;
-				$safety = 0;
-				//Do not let the stat go over 31
-				//2018 08 30 - ealm - Instead of rerolling, let's just use the next stat.  I guess this encourages high stats to be adjacent though?
-				while(getMonsterByte($i, 12 + $slot) >= 31){
-					$slot = ($slot + 1) % 6;
-					$safety++;
-					if($safety >= 6) break;
-				}
-				if($safety >= 6) break;
-				if (getMonsterByte($i, 12 + $slot) < 31)
-				{
-					setMonsterByte($i, 12 + $slot, getMonsterByte($i, 12 + $slot) + 1);
-					$total_stats--;
-				}
-			}
-		} elseif ($Flags["Growth"] == "Shuffle") {
-			$options = [];
-			for ($j = 0; $j < 6; $j++)
-			{
-				$options[] = getMonsterByte($i, 12 + $j);
-			}
-			for ($j = 0; $j < 6; $j++)
-			{
-				$chosen_offset = Random() % count($options);
-				$chosen = array_splice($options, $chosen_offset, 1);
-				setMonsterByte($i, 12 + $j, $chosen[0]);
-			}
-		}
-		
-		//If we're in Genius Mode, all monsters get 31 int growth
-		if($Flags["GeniusMode"] == "Yes"){
-			setMonsterByte($i, 12 + 5, 31);
-		}
-
-	}
-
-	return true;
-}
-
-
-
-function ShuffleMonsterResistances()
-{
-	global $Flags;
-	
-	global $romData;
-	global $monster_count;
-	
-	for ($i = 0; $i < $monster_count; $i++)
-	{
-		if ($Flags["Resistance"] == "Redistribute")
-		{
-			//Repeat for resistances.  There are 27 of these...
-			$total_resistances = 0;
-			for ($j = 0; $j < 27; $j++)
-			{
-				$total_resistances += getMonsterByte($i, 18 + $j);
-				setMonsterByte($i, 18 + $j, 0);
-			}
-			if($total_resistances > 27*3) $total_resistances = 27*3;
-			while ($total_resistances > 0)
-			{
-				$slot = Random() % 27;
-				//2019 03 11 - ealm - Initialize this variable idiot
-				$safety = 0;
-				//2018 08 30 - ealm - Instead of rerolling, let's just use the next stat.  I guess this encourages high stats to be adjacent though?
-				while(getMonsterByte($i, 18 + $slot) >= 3){
-					$slot = ($slot + 1) % 27;
-					$safety++;
-					if($safety >= 27) break;
-				}
-				if($safety >= 27) break;
-				//Do not let the stat go over 3
-				if (getMonsterByte($i, 18 + $slot) < 3)
-				{
-					setMonsterByte($i, 18 + $slot, getMonsterByte($i, 18 + $slot) + 1);
-					$total_resistances--;
-				}
-			}
-		} elseif ($Flags["Resistance"] == "Shuffle") {
-			$options = [];
-			for ($j = 0; $j < 27; $j++)
-			{
-				$options[] = getMonsterByte($i, 18 + $j);
-			}
-			for ($j = 0; $j < 27; $j++)
-			{
-				$chosen_offset = Random() % count($options);
-				$chosen = array_splice($options, $chosen_offset, 1);
-				setMonsterByte($i, 18 + $j, $chosen[0]);
-			}
-		}
-	}
-
-	return true;
-}
-
-
-function ShuffleMonsterSkills()
-{
-	global $Flags;
-	global $romData;
-	global $monster_count;
-	global $magicValues;
-	global $SkillIDsByName;
-
-	$tier_one_skills = array( 1, 4, 7, 10, 13, 16, 19, 21, 22, 25, 27, 30, 32, 33, 34, 35, 36, 37, 39, 41, 43, 45, 46, 47, 49, 51, 52, 53, 54, 56, 57, 58, 60, 61, 62, 63, 64, 68, 72, 74, 75, 76, 78, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 137, 138, 139, 141, 143, 144, 145, 146, 147, 148, 149, 150, 151, 153, 155, 156, 157, 158, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169 );
-
-	// Remove BeDragon unless we've specifically asked to keep it.
-	if ($Flags["Skills"] != "Random With BeDragon") {
-		$index = array_search($SkillIDsByName["BeDragon"], $tier_one_skills);
-		if ($index !== false) {
-			unset($tier_one_skills[$index]);
-		}
-	}
-
-	if ($Flags["Skills"] == "Random, No Healing") {
-		foreach ($magicValues["Healing Skills"] as $skillId) {
-			$index = array_search($skillId, $tier_one_skills);
-			if ($index !== false) {
-				unset($tier_one_skills[$index]);
-			}
-		}
-	}
-  
-	for ($i = 0; $i < $monster_count; $i++)
-	{
-		if ($Flags["Skills"] != "None")
-		{
-			//Randomize skills!  Pick three of these.
-			$skill1 = Random() % count($tier_one_skills);
-			$skill2 = Random() % count($tier_one_skills);
-			while ($skill2 == $skill1)
-			{
-				$skill2 = Random() % count($tier_one_skills);
-			}
-			$skill3 = Random() % count($tier_one_skills);
-			while ($skill3 == $skill1 || $skill3 == $skill2)
-			{
-				$skill3 = Random() % count($tier_one_skills);
-			}
-			setMonsterByte($i, 8, $tier_one_skills[$skill1]);
-			setMonsterByte($i, 9, $tier_one_skills[$skill2]);
-			setMonsterByte($i, 10, $tier_one_skills[$skill3]);
-		}
-	}
-
-	return true;
-}
-
-function isArenaEncounter($i) {
-	// Kid class: 130-136
-	// Class C: 137-145
-	// Class B: 146-151?
-	// Class A: 152-160
-	// Class S: 161-167
-	// KingLeo/Azurile/Divinegon: 168-170
-	// Post-game arena randos?: 331-375
-	if ($i >= 130 && $i <= 170) { return 1; }
-	if ($i >= 331 && $i <= 375) { return 1; }
-	return 0;
-}
-
-function isBossEncounter($i)
-{
-	// Treat all encounters that have a 0% recruitment rate as bosses. This includes arena fights.
-	if (getEncounterByte($i, 8) == 7) { return 1; }
-
-	// There are three boss encounters that apparently allow normal recruitment mechanics from the fight.
-	switch($i){
-		case 6: //Oasis CurseLamp
-		case 399: //Pirate KingSquid
-		case 426: //Sky MadCondor
-			return 1;
-	}
-  return 0;
-}
-
-function isBossRecruit() {
-	// Some of the bosses can be recruited through storyline means, but their recruitable stats
-	// are stored in a different encounter than the actual boss fight.
-	// TODO: Missing some?
-	switch($i){
-		case 386: //Oasis Beavern
-		case 7: //Oasis CurseLamp
-		case 26: //Pirate Hoodsquid
-		case 427: //Sky MadCondor
-		case 430: //Pirate KingSquid
-			return 1;
-	}
-  return 0;
-}
-
-// Other encounter notes:
-// Randomly found mimics?: 117-124 (levels 1/5/10/15/20/25/30/35)
-// MedalKing egg rewards: 377-382 (MadCat, HornBeet, Skydragon, Octogon, Servant, Darck)
-// Wandering monster master randos?: 171-330
-
-function ShuffleEncounters()
-{
-	global $Flags;
-	
-	global $romData;
-	global $ValidMonsterIDs;
-	global $ValidMonsterGrowthIndecies;
-	global $MonsterIDsByName;
-	global $encounter_count;
-	global $magicValues;
-	
-	//Code patch: Reduce level of SpikyBoys in Oasis to 1 so that they level faster
-	$romData[0xD00CC] = chr(0x01);
-
-	for ($i = 0; $i < $encounter_count; $i++)
-	{
-		//Which monster is this?
-		if($i == 0 && $Flags["StartingMonster"] != 0){
-			//Allow the starting monster to be selectable
-			$monsterid = $Flags["StartingMonster"];
-		}
-		elseif($i == $magicValues["HoodSquid Drop Encounter"]){
-			//Special case: The hoodsquid needs to be a water-type. (0x13C - 0x15B, 32 monsters)
-			//TODO: Yeti Mode won't put a yeti here ):
-			$monsterid = Random() % 32 + 0x13C;
-		}else{
-			$monsterid = $ValidMonsterIDs[Random() % count($ValidMonsterIDs)];
-		}
-		$MonsterGrowthIndex = array_search($monsterid,$ValidMonsterGrowthIndecies);
-		
-		//Should probably choose monster independently of the rest of this.
-		if($Flags["Encounters"] == "Poorly") //Previously "Based On Growth"
-		{
-			
-			//Need to ensure Army Ant/Madgopher are obtainable before Ice.
-			if(getEncounterWord($i, 0) != $MonsterIDsByName["MadGopher"] && getEncounterWord($i, 0) != $MonsterIDsByName["ArmyAnt"]){
-				setEncounterWord($i, 0, $monsterid);
-			}
-			
-			$is_boss_I_think = isBossEncounter($i);
-			
-			//Add up the monster's GROWTH values
-			$total_growth_stats = 0;
-			for ($j = 0; $j < 6; $j++)
-			{
-				if($j == 0 && $is_boss_I_think) continue;
-				$total_growth_stats += getMonsterByte($MonsterGrowthIndex, 12 + $j);
-			}
-			
-			//Add up the monster's BASE STATS
-			$total_stats = 0;
-			for ($j = 0; $j < 6; $j++)
-			{
-				if($j == 0 && $is_boss_I_think) continue;
-				$total_stats += getEncounterWord($i, 10 + $j * 2);
-			}
-			//Double the base stats for our starting monster, because Slash is a little weak at level one.
-			if($i == 0) $total_stats *= 2;
-			
-			//Take the percentage of the GROWTH allocated to each stat and multiply by the total BASE
-			for ($j = 0; $j < 6; $j++)
-			{
-				//TODO: Test stat scaling Kappa
-				if($j == 0 && $is_boss_I_think){
-					//2018 08 30 - ealm - TODO: Why is this continue here?
-					//continue;
-					//We're handling boss HP differently.  Basically, we're not changing it.  However...
-					//2018 06 25 - ealm - Adding base stat scaling flags
-					$new_stat = getEncounterWord($i, 10 + $j * 2);
-					
-					//I actually don't want HP to scale the same as other stats.  Let's "soften" the effect of the scalar if it's over 100%.
-					$scalar = $Flags["BossScaling"]/100;
-					//Turns 500/400/300/200 into 300/250/200/150.
-					if($scalar > 1) $scalar = ($scalar - 1) / 2 + 1;
-										
-					setEncounterWord($i, 10 + $j * 2, $new_stat);
-				}else{
-					//2018 06 25 - ealm - Renaming this variable for clarity.  This is the growth value for this stat.
-					$stat_growth = getMonsterByte($MonsterGrowthIndex, 12 + $j);
-					//Divide by total growth stats to get this stat's "share" ratio, multiply by total base stats
-					$new_stat = floor($stat_growth * $total_stats / $total_growth_stats);
-					
-					//2018 06 25 - ealm - Adding base stat scaling flags
-					$scalar = 1;
-					if($is_boss_I_think) $scalar = $Flags["BossScaling"]/100;
-					else $scalar = $Flags["StatScaling"]/100;
-					$new_stat *= $scalar;
-					
-					
-					
-					//Everyone gets minimum 10 HP and 5 str, agi, def, int.  (MP can be zero)
-					//If this is our starting monster (Slash), double the baseline for all of those.
-					if($j == 0){
-						if($new_stat < 10 * (($i == 0) ? 2 : 1)) $new_stat = 10 * (($i == 0) ? 2 : 1);
-					}elseif(($j == 2) or ($j == 3) or ($j == 4) or ($j == 5)){
-						if($new_stat < 5 * (($i == 0) ? 2 : 1)) $new_stat = 5 * (($i == 0) ? 2 : 1);
-					}
-					//2018 06 25 - ealm - Why wasn't this already in here?  Please cap stats (except boss HP, handled above) at 999...
-					if($new_stat > 999) $new_stat = 999;
-					
-					setEncounterWord($i, 10 + $j * 2, $new_stat);
-				}
-			}
-		}
-		//Ramp up early EXP gains.
-		$exp = getEncounterWord($i, 6);
-		if($exp < 20) {
-			setEncounterWord($i, 6, $exp*2.5);
-		} elseif($exp < 40) {
-			setEncounterWord($i, 6, $exp*2);
-		} elseif($exp < 100) {
-			setEncounterWord($i, 6, $exp*1.5);
-		}
-		$global_exp_scalar = $Flags["EXPScaling"]/100;
-		$exp = getEncounterWord($i, 6) * $global_exp_scalar;
-		// EXP is 2 bytes. Don't overflow and end up with a small EXP amount.
-		if ($exp > 0xFFFF) {
-			$exp = 0xFFFF;
-		}
-		setEncounterWord($i, 6, $exp);
-		
-		//If we're in Genius Mode, all wild monsters get 999 int
-		if($Flags["GeniusMode"] == "On"){
-			setEncounterWord($i, 10 + 5*2, 999);
-		}
-
-		
-		//Now, let's teach random encounters their skills.  We'll give them the three they're supposed to learn, plus a bonus skill, and let it level up appropriately.
-		$lv  = getEncounterByte($i, 9);
-		$hp  = getEncounterWord($i, 10);
-		$mp  = getEncounterWord($i, 12);
-		$atk = getEncounterWord($i, 14);
-		$def = getEncounterWord($i, 16);
-		$agl = getEncounterWord($i, 18);
-		$int = getEncounterWord($i, 20);
-		
-		for($j = 0; $j < 4; $j++){
-			//Loop through all three skills the monster should learn, plus a BONUS SKILL
-			if($j <> 3){
-				$skill = getMonsterByte($MonsterGrowthIndex, 8 + $j);
-			}else{
-				$skill = Random() % 169 + 1;
-				//Re-roll until this isn't the same skill as the three it innately learns.
-				while($skill == getMonsterByte($MonsterGrowthIndex, 8) ||
-					  $skill == getMonsterByte($MonsterGrowthIndex, 9) ||
-					  $skill == getMonsterByte($MonsterGrowthIndex, 10))
-					  {
-					$skill = Random() % 169 + 1;
-				}
-			}
-			
-			//This variable is the "return value" from our loop.
-			$return_skill = 0xFF; //0xFF is "no skill".
-			while(true){
-				//"Skill" is the ID of the skill we're trying to learn.
-				$skill_qry = "select * from dragonwarriormonsters2_skills where id = ".$skill." and lv <= ".$lv." and hp <= ".$hp." and mp <= ".$mp." and atk <= ".$atk." and def <= ".$def." and agl <= ".$agl." and `int` <= ".$int;
-				execute($skill_qry);
-				$result = get();
-				if(count($result) !== 0){
-					//If we can learn this skill, plug it into return_skill
-					$return_skill = $result["id"];
-					$skill = $result["SUCCESSOR"];
-					//echo 'Evolve into '.$result['Name'].': '.$lv.', '.$hp.', '.$mp.', '.$atk.', '.$def.', '.$agl.', '.$int.'<br>';
-					if($skill == 0){
-						//Break if there is no next skill to look up.
-						break;
-					}
-				}else{
-					//echo 'No evolution: '.$lv.', '.$hp.', '.$mp.', '.$atk.', '.$def.', '.$agl.', '.$int.'<br>';
-					//Break if we don't qualify for the current skill.
-					break;
-				}
-			}
-			setEncounterByte($i, 2 + $j, $return_skill);
-		}
-		//Hoodsquid should always know LureDance as its fourth move
-		if($i == $magicValues["HoodSquid Drop Encounter"]){
-			setEncounterByte($i, 4, $SkillIDsByName["LureDance"]);
-		}
-		
-		//Swap empty moves to the back.  Just gonna "brute force" a bubble sort; could be more efficient but it's nine swaps max so whatever.
-		for($swapCounter = 0; $swapCounter < 3; $swapCounter++) {
-			for($j = 2; $j <= 4; $j++){
-				if(getEncounterByte($i, $j) == 0xFF){
-					swapEncounterBytes($i, $j, $j + 1);
-				}
-			}
-		}
-		//END ENCOUNTER SKILLS
-	}
-
-	// Old to-do list:
-	//TODO: If it needs it, make starting monster have a minimum of 10 on each stat (15-20 for HP/MP?)
-	//TODO: Pick from three monsters instead of just selecting one
-	//TODO: Distribute stats on monsters by deleveling them and then leveling them back up
-	//TODO: Does everyone actually have 31 int growth or did I forget that?
-	//TODO: All enemy monsters seem to target back monster?
-	//TODO: Why did Putrepup in Ice world hit so hard?
-	//TODO: Sky world enemies hurt!!!
-	//TODO: Mudo's HP was randomized and it hurt WutFace
-	//TODO: Make WLD zero for all random encounters
-	//TODO: Why is this where the TODOs are?
-	//TODO: Maybe make it easier for random encounters to learn skills, so that early monsters know things? (Only use level requirement?)
-	
-	return true;
-}
-
-function CodePatches(){
-	global $romData;
-	//Wherever "clear water" is mentioned, write "tonic" over the word "water"
-	//TODO: Find more of these; Bizhawk's text search is glitchy.  Do a text dump?
-	WriteText(0x266624, "tonic");
-	WriteText(0x0A107B, "tonic");
-	
-	//TODO: This code to randomize your team name in the arena is glitchy...  If I write too many (or too few?) characters over existing text, the text box can glitch out.  Might need to figure out what the extra characters are between messages...
-	//This is an array of every instance of "Master ____ and the (Team Name)!", which changes with every rank.
-	//These aren't hex values because my text dump is broken into 100-character lines.  Don't judge me.
-	/*
-	$Masters = [[155640,30],[155676,29],[155711,28],[155745,30],[155781,27],[155885,32],[156338,30],[156441,30],[156549,29],[156656,31],[156766,28],[156866,31],[156976,32]];
-	$Titles = ["Master","Captain","Commander","Mistress","Lord","Lady","Sir","King","Queen","Princess","Prince","Duke","Duchess","Speedrunner","WR Holder","Lieutenant","Corporal","Officer","Admiral","Pirate Lord"];
-	$Teams = ["Team","Sweaty Yetis","Alefgard Alliance","Dream Team","Good Monsters","Slime Time","Fun Friends","Berserkers","Random Monsters","Buttermilk Ranch"];
-	for($i = 0; $i < count($Masters); $i++){
-		//We've got 31 characters to work with, including 0xf900 which'll be the player's name
-		//"(Title) ____ and the (Team)" = 12 characters + title/team names (Player name is two characters)
-		//(Note to self:  I gotta make sure the longest title and the shortest team name don't exceed 20 characters)
-		$title = Random() % count($Titles);
-		$team = Random() % count($Teams);
-		//Roll team names until we get one that's short enough.
-		while(strlen($Titles[$title]) + strlen($Teams[$team]) + 12 > $Masters[$i][1]){
-			$team = Random() % count($Teams);
-		}
-		$_team_with_padding = $Teams[$team];
-		while(strlen($Titles[$title]) + strlen($_team_with_padding) + 12 < $Masters[$i][1]){
-			$_team_with_padding .= ' ';
-		}
-		
-		WriteText($Masters[$i][0], $Titles[$title].' ');
-		$romData[$Masters[$i][0]+strlen($Titles[$title])+1] = chr(0xF9);
-		$romData[$Masters[$i][0]+strlen($Titles[$title])+2] = chr(0x00);
-		WriteText($Masters[$i][0]+strlen($Titles[$title])+3, ' and the '.$_team_with_padding);
-	}
-	*/
-
-	//Patch to open up the monster breeder's back door early (As soon as you get Slash)
-	//0x3FAD-0x3FFF are no-ops, so we'll write our new code in that range.
-	
-	//First, move these three addresses that we're about to replace with the JP command into the new code block
-	$romData[0x3FAD] = $romData[0x186C];
-	$romData[0x3FAE] = $romData[0x186D];
-	$romData[0x3FAF] = $romData[0x186E];
-	//$romData[0x3FB0] = $romData[0x186F];
-	
-	$romData[0x186C] = chr(0x00); 	//	nop
-	$romData[0x186D] = chr(0xC3); 	//	JP 0x3FAD	Jump to 0x3FAD
-	$romData[0x186E] = chr(0xAD);	//
-	$romData[0x186F] = chr(0x3F);	//
-	
-	//Now, add my own code...
-	$romData[0x3FB1] = chr(0x3E);	//LD a,FCh		Load 0xFC into register a
-	$romData[0x3FB2] = chr(0xFC);	//
-	$romData[0x3FB3] = chr(0x5F);	//LD e,a		Load register A into register E
-	
-	//$romData[0x3FB4] = chr(0xCB);	//SLA C			(Turn C from 4 to 8 with a left shift, two-byte command)
-	//$romData[0x3FB5] = chr(0x21);	//
-	$romData[0x3FB4] = chr(0x0E);	//SLA C			(Load the number 8 into register c)
-	$romData[0x3FB5] = chr(0x08);	//
-	$romData[0x3FB6] = chr(0x1A);	//ld a,(de)
-	$romData[0x3FB7] = chr(0xB1);	//or c
-	$romData[0x3FB8] = chr(0x12);	//ld (de),a
-	
-	$romData[0x3FB9] = chr(0xC9);	//	RET 		End subroutine (Instruction previously at 0x186F)
-	//TODO: Is it possible to get rid of all of the people in front of the monster breeder?
-}
-
-function LocateBosses(){
-	//I haven't been citing my sources for all of the information I'm pulling from Gamefaqs since it's in the ROM data anyway...
-	//...but this data doesn't seem to be entirely accurate thecowLUL
-	//boss stats source: https://gamefaqs.gamespot.com/gbc/525414-dragon-warrior-monsters-2-cobis-journey/faqs/14383
-	
-	global $romData;
-	global $encounter_count;
-
-	//Really shoulda put this in the database...
-	$boss_stats = array(
-		array("Oasis Beavern",5,98,16,20,8,36,120),
-		array("Oasis CurseLamp",5,220,8,27,10,44,65),
-		array("K-1 Babble",6,36,9,14,13,32,123),
-		array("K-1 PearlGel",5,15,12,17,35,51,96),
-		array("K-2 SpikyBoy",6,19,42,17,31,34,96),
-		array("K-2 Pixy",7,29,24,20,14,54,79),
-		array("K-2 Dracky",7,40,22,14,12,70,96),
-		array("K-3 MadRaven",9,41,55,16,30,84,163),
-		array("K-3 Kitehawk",9,46,35,26,21,42,146),
-		array("K-3 MadRaven",9,41,55,16,30,84,163),
-		array("Pirate Hoodsquid",12,350,100,78,39,85,180),
-		array("Pirate Boneslave",13,72,58,58,45,49,99),
-		array("Pirate CaptDead",13,500,105,66,48,50,144),
-		array("Pirate KingSquid",38,2500,140,227,147,189,73),
-		array("Ice Bombcrag",18,650,10,80,80,20,140),
-		array("Ice AgDevil",18,550,50,90,59,109,216),
-		array("Ice Puppetor",22,400,82,85,60,48,120),
-		array("Ice Goathorn",27,850,89,95,65,92,165),
-		array("Ice ArcDemon",27,210,89,105,71,84,211),
-		array("Ice Goathorn 2",27,330,89,95,65,92,223),
-		array("Sky MadCondor",30,600,119,131,90,110,159),
-		array("Sky Skeletor",37,226,108,160,148,140,140),
-		array("Sky Niterich",36,1500,106,177,149,110,510),
-		array("Sky Metabble",37,20,368,95,999,670,522),
-		array("Sky EvilArmor",38,450,140,185,150,138,280),
-		array("Sky Mudou",38,3000,413,225,160,156,150),
-		array("Limbo GigaDraco",42,1000,245,277,180,120,100),
-		array("Limbo Centasaur",40,900,83,220,150,250,250),
-		array("Limbo Garudian",41,800,88,206,160,251,320),
-		array("Limbo Darck",44,4000,235,350,220,160,210),
-		array("Butch Butch",0,755,322,547,238,644,500),
-		array("Butch Pumpoise",40,800,388,238,255,330,500),
-		array("Butch Drygon",40,480,210,432,677,270,500),
-		array("Kameha50 MimeSlime",43,370,690,258,320,350,520),
-		array("Kameha50 Tonguella",45,400,320,338,310,284,430),
-		array("Kameha50 Golem",48,520,315,410,301,189,540),
-		array("Kameha150 MetalKing",50,200,950,310,780,840,700),
-		array("Kameha150 KingLeo",55,1200,490,370,480,460,680),
-		array("Kameha150 GoldGolem",50,900,590,430,600,370,700),
-		array("Terry GreatDrak",52,840,330,375,400,299,300),
-		array("Terry Watabou",40,610,467,320,380,178,150),
-		array("Terry Durran",53,1000,470,420,430,326,450),
-		array("Elf Arrowdog",40,500,210,237,288,249,501),
-		array("Elf AgDevil",47,1500,580,346,310,267,700),
-		array("Power WindBeast",40,900,127,236,168,489,0),
-		array("Power MadGoose",40,580,387,201,203,348,0),
-		array("Power WhaleMage",42,750,590,199,228,247,0),
-		array("Power SeaHorse",37,480,180,191,253,222,0),
-		array("Power Octoreach",39,700,140,238,174,247,0),
-		array("Power IceMan",45,1200,500,345,279,210,0),
-		array("Power Shadow",40,600,300,235,240,170,0),
-		array("Power BigEye",37,750,280,358,242,198,0),
-		array("Power Balzak",45,2000,398,375,257,220,0),
-		array("Power RotRaven",35,680,340,160,249,340,0),
-		array("Power Jamirus",48,1400,750,247,349,410,0),
-		array("Power SkyDragon",46,990,700,349,279,243,0),
-		array("Power Gremlin",40,600,450,201,232,218,0),
-		array("Traveler Pixy",52,680,400,357,374,443,700),
-		array("Traveler Copycat",55,750,300,310,520,380,999),
-		array("Traveler StoneMan",52,1300,360,458,268,247,700),
-		array("Traveler WhipBird",53,950,830,334,374,438,700),
-		array("Traveler MetalKing",48,600,200,410,859,320,160),
-		array("Traveler RainHawk",99,3500,520,710,540,250,300),
-		array("Traveler Coatol",46,1500,210,600,600,300,255),
-		array("Baffle Crestpent",42,780,700,279,300,299,480),
-		array("Baffle SpotKing",45,1300,550,378,214,308,800),
-		array("Baffle Gulpple",45,840,590,289,299,387,0),
-		array("Baffle FairyDrak",40,700,340,312,298,273,0),
-		array("Baffle DuckKite",42,900,470,296,330,400,0),
-		array("Baffle FunkyBird",42,890,500,301,267,430,0),
-		array("Baffle FunkyBird",42,890,500,301,267,430,0),
-		array("Baffle Slurperon",39,790,400,277,259,302,0),
-		array("Baffle DanceVegi",37,680,430,259,249,240,0),
-		array("Baffle MadPlant",39,880,380,243,350,289,510),
-		array("Baffle Orc",43,900,350,346,289,190,770,6000),
-		array("Baffle PutrePup",41,700,280,330,328,232,450),
-		array("Baffle Devipine",43,840,400,300,358,222,660),
-		array("Baffle Anemon",44,900,370,248,388,289,580),
-		array("Baffle HerbMan",46,1300,700,341,273,370,900),
-		array("Soul KingSlime",50,4000,280,329,289,334,800),
-		array("Soul Coatol",48,640,800,289,410,329,800),
-		array("Soul FangSlime",44,700,378,333,278,397,590),
-		array("Soul Grizzly",45,870,360,372,245,379,288),
-		array("Soul BeastNite",45,1200,420,358,343,201,600),
-		array("Soul Slimeborg",44,650,790,217,327,418,740),
-		array("Soul Unicorn",45,720,680,316,322,319,800),
-		array("Soul SuperTen",44,600,380,312,298,417,600),
-		array("Soul Slime",47,600,500,387,299,387,700),
-		array("Soul RockSlime",40,680,230,265,362,265,450),
-		array("Soul Metabble",40,289,700,146,798,688,770),
-		array("Soul Gorago",55,990,780,432,378,688,800)
-	);
-	
-	foreach($boss_stats as $boss){
-		$foundit = 0;
-		for ($i = 0; $i < $encounter_count; $i++){
-			$lv  = getEncounterByte($i, 9);
-			$hp  = getEncounterWord($i, 10);
-			$mp  = getEncounterWord($i, 12);
-			$atk = getEncounterWord($i, 14);
-			$def = getEncounterWord($i, 16);
-			$agl = getEncounterWord($i, 18);
-			$int = getEncounterWord($i, 20);
-			
-			$pts = 0;
-			if($boss[1] == $lv ) $pts++;
-			if($boss[2] == $hp ) $pts++;
-			if($boss[3] == $mp ) $pts++;
-			if($boss[4] == $atk) $pts++;
-			if($boss[5] == $def) $pts++;
-			if($boss[6] == $agl) $pts++;
-			if($boss[7] == $int) $pts++;
-			
-			if($pts > 2){
-				$foundit = 1;
-				echo $boss[0].' located: '.$i.' ('.$pts.')<br>';
-			}
-		}
-		if($foundit == 0){
-			echo $boss[0].' NOT located!<br>';
-		}
-	}
-}
-
-function flagRadioButtons($name, $label, $options) {
-	global $Flags;
-?>
-	<div class="row">
-		<div class="col-sm"><?php echo $label ?></div>
-<?php
-		foreach ($options as $option) {
-			$id = strtolower($name) . '_' . strtolower($option["value"]);
-?>
-			<div class="col-sm"><input type="radio" name="<?php echo $name ?>" value="<?php echo $option["value"] ?>" id="<?php echo $id ?>" <?php echo $Flags[$name] == $option["value"] ? 'checked' : '' ?> /> <label for="<?php echo $id ?>" title="<?php echo $option["description"] ?>"><?php echo $option["label"] ? $option["label"] : $option["value"] ?></label></div>
-<?php
-		}
-?>
-	</div>
-<?php
-}
-
-DWM2R();
+require_once("include/db.php");
+require_once("include/globals.php");
+require_once("include/random.php");
+require_once("include/rom.php");
+require_once("include/rom_dump.php");
+require_once("include/rom_hacks.php");
+require_once("include/ui.php");
+
+main();
 
 ?>
 <html>
 <head>
 	<script type="text/javascript" src="/Library/jquery-3.0.0.js"></script>
 	<script type="text/javascript" src="/Library/bootstrap-4.0.0/js/bootstrap.js"></script>
+	<script type="text/javascript" src="dwm2rando.js"></script>
 	<link rel="stylesheet" href="/Library/bootstrap-4.0.0/css/bootstrap.css" />
+	<link rel="Stylesheet" href="dwm2rando.css" />
 </head>
 <body>
 
-<form action="index.php" method="POST" enctype="multipart/form-data">
-<div class="container-fluid" style="max-width:800px;margin:0px auto;">
-	<div class="row">
-		<div class="col-sm" style="text-align:center"><h1>Dragon Warrior Monsters 2 Randomizer!</h1></div>
-	</div>
-	<div class="row">
-		<div class="col-sm"><label for="flags_input">Flags:</label></div>
-		<div class="col-sm"><input type="text" name="Flags" value="" id="flags_input" size=15 /></div>
-		<div class="col-sm"><label for="seed_input">Seed:</label></div>
-		<div class="col-sm"><input type="text" name="Seed" value="<?php echo $Flags['Seed'] ?>" id="seed_input" size=10 /></div>
-		<div class="col-sm"><input type="button" name="NewSeedBtn" value="New Seed" id="NewSeedBtn" /></div>
-	</div>
-	<div class="row">
-		<div class="col-sm"><label for="input_file">Input File:</label></div>
-		<div class="col-sm"><input type="file" name="InputFile" value="" id="input_file" size=50 /></div>
-	</div>
-	<div class="row">
-		<div class="col-sm">Starting Monster</div>
-		<div class="col-sm">
-			<select name="StartingMonster">
-				<option value="0">Random</option>
-				<?php
-					$monster_list_query = "SELECT * FROM dragonwarriormonsters2 order by id asc";
-					execute($monster_list_query);
-					while($monster = get()){
-				?>
-				<option value="<?php echo $monster["id"]; ?>" <?php echo $Flags['StartingMonster'] == $monster["id"] ? 'selected' : '' ?>><?php echo $monster["name"]; ?></option>
-				<?php
-					}
-				?>
-			</select>
-		</div>
-	</div>
-<?php
-	foreach ($FlagSettings as $name => $settings) {
-		flagRadioButtons($name, $settings["label"], $settings["options"]);
-	}
-?>
-	<div class="row">
-		<div class="col-sm" style="text-align:center"><input type="Submit" name="Submit" value="Randomize!"></div>
-	</div>
-</div>
-</form>
-<div class="container-fluid" style="max-width:800px;margin:0px auto;">
-	<div class="row">
-		This randomizer will work with both Cobi's Journey or Tara's Adventure, but the resulting rom will vary slightly based on which one you use.  Note that this randomizer is still in beta, and consequently has a few known bugs and isn't quite feature-complete.
-	</div>
-	<div class="row">
-		For more details on how this randomizer works, planned features and changelog, and the source code/data, checkout the ReadMe on GitHub: <a href="https://github.com/TheCowness/DWM2Randomizer">https://github.com/TheCowness/DWM2Randomizer</a>
-	</div>
-</div>
-
-<style type="text/css">
-	form > .container-fluid{
-		border-left:2px solid #ccc;
-		border-top:2px solid #ccc;
-	}
-	form > .container-fluid > .row{
-		border-right:2px solid #ccc;
-		border-bottom:2px solid #ccc;
-	}
-
-	form .row b {
-		color: #F33;
-	}
-</style>
-
-<script type="text/javascript">
-$(document).ready(function(){
-	//Set a random seed, if we didn't just generate one.
-	if($("#seed_input").val() == '0'){
-		NewSeed();
-	}
-	$("#NewSeedBtn").click(function(){NewSeed();});
-	
-	//When any flag-field is changed, update the flag string
-	$("input[type=radio]").change(function(){
-		RefreshFlagString();
-	});
-	$("select").change(function(){
-		RefreshFlagString();
-	});
-	//When the flag string is changed, update all fields
-	$("#flags_input").change(function(){
-		SetFlagsFromString();
-	});
-	//Initialize the flag string
-	RefreshFlagString();
-});
-function NewSeed(){
-	$("#seed_input").val(Math.floor(Math.random()*268435455));
-}
-
-
-//Number of 6-bit values required to represent an HTML "select" value.
-select_max_size = 2;
-function RefreshFlagString(){
-	//This function will serialize the important form flags into a text string so that they can be shared for races.
-	//Because JS integers are basically 32-bit, we'll have to utilize an array.
-	//Let's have each array value hold a six-bit integer (0-63), which will ultimately be represented in text by a single character
-	//I realized later that I basically reinvented base-64 encoding here...
-	flags = [];
-	flag_ctr = 0;
-	flag_chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_-';
-	//Start with all radio buttons.  Six radio buttons can be stored in one 6-bit array value
-	$("input[type=radio]").each(function(){
-		if(flag_ctr % 6 == 0){
-			//Initialize the next array index
-			flags.push(0);
-		}
-		if($(this).prop("checked")){
-			flags[Math.floor(flag_ctr / 6)] += 2**(flag_ctr % 6);
-		}else{
-			flags[Math.floor(flag_ctr / 6)] += 0;
-		}
-		flag_ctr++;
-	});
-	//Now do selects.  Assume all select values are integers less than (64 ^ select_max_size)
-	$("select").each(function(){
-		for(i = 0; i < select_max_size; i++){
-			flags.push(Math.floor($(this).val() / (64 ** i)) % 64);
-		}
-	});
-	
-	//Convert our array into a "base 64" string and output to the Flags field.
-	$("#flags_input").val('');
-	for(i = 0; i < flags.length; i++){
-		$("#flags_input").val($("#flags_input").val()+flag_chars.charAt(flags[i]));
-	}
-}
-function SetFlagsFromString(){
-	//This will be the opposite of RefreshFlagString - turn on flags based on the Flags string.
-	//TODO: Need a check to make sure the flags string is long enough.  Calculate length based on number of form elements?
-	flag_string = $("#flags_input").val();
-	flags = [];
-	flag_ctr = 0;
-	flag_chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_-';
-	//Convert from alphanumeric to the array we used above
-	for(i = 0; i < flag_string.length; i++){
-		flags.push(flag_chars.indexOf(flag_string.charAt(i)));
-	}
-	//Start with all radio buttons
-	$("input[type=radio]").each(function(){
-		if((flags[Math.floor(flag_ctr / 6)] & 2**(flag_ctr%6)) != 0){
-			$(this).prop("checked","checked");
-		}else{
-			$(this).removeProp("checked");
-		}
-		flag_ctr++;
-	});
-	//Continue with Selects.
-	$("select").each(function(){
-		_newval = 0;
-		for(i = 0; i < select_max_size; i++){
-			_newval += flags[flag_ctr / 6] * 64 ** i;
-			flag_ctr += 6;
-		}
-		$(this).val(_newval);
-	});
-}
-</script>
+<?php drawForm(); ?>
 
 </body>
 </html>
